V1.2

+++ functions/index.js
@@ -1,11 +1,10 @@
 process.env.TZ = "America/New_York";
 
-const { log, info, debug, warn, error, write } = require("firebase-functions/logger");
-const { onCall, HttpsError, onRequest } = require("firebase-functions/v2/https");
 const admin = require("firebase-admin");
-const { getFirestore } = require("firebase-admin/firestore");
+const { log, info, debug, warn, error, write } = require("firebase-functions/logger");
+const { onCall } = require("firebase-functions/v2/https");
+const { getFirestore, Timestamp } = require("firebase-admin/firestore");
 const { onSchedule } = require("firebase-functions/v2/scheduler");
-const fs = require("fs");
 const PDFDocument = require("pdfkit");
 const { getTimezoneOffset } = require("date-fns-tz");
 
@@ -21,7 +20,7 @@ const send = async (options) => {
 		let messageId = await sendMail(options);
 		return messageId;
 	} catch (error) {
-		error("There was an error connecting to Gmail: ", { errorMsg: error });
+		console.log("There was an error connecting to Gmail: ", { errorMsg: error });
 	}
 };
 
@@ -30,6 +29,7 @@ exports.sendEmail = onCall(
 		enforceAppCheck: true, // Reject requests with missing or invalid App Check tokens.
 	},
 	(request) => {
+		if (request.auth === null) return;
 		let messageOptions = emailOptions;
 		messageOptions.to = request.data.email;
 		messageOptions.subject = request.data.subject;
@@ -43,7 +43,8 @@ exports.bugReport = onCall(
 	{
 		enforceAppCheck: true,
 	},
-	(request) => {
+	async (request) => {
+		if (request.auth === null) return;
 		let messageOptions = emailOptions;
 		messageOptions.to = "bailey.tuckman@westtown.edu";
 		messageOptions.cc = "polaris@westtown.edu";
@@ -55,7 +56,26 @@ exports.bugReport = onCall(
 			day: "numeric",
 		})} there was a bug reported on ${request.data.page} by ${
 			request.data.email ? request.data.email : "anonymous"
-		}.\n\nDescription:\n${request.data.description}\n\nSteps to Reproduce: ${request.data.repro}`;
+		}.\n\nDescription:\n${request.data.description}\n\nSteps to Reproduce: ${
+			request.data.repro
+		}\n\nCheck this bug report here: https://polaris-60dce.web.app/admin/settings.html`;
+
+		let currentNum = await db.collection("bugReports").count().get();
+		db.collection("bugReports")
+			.doc(`${currentNum.data().count}`)
+			.set({
+				date: Timestamp.fromDate(new Date()),
+				description: request.data.description,
+				page: request.data.page,
+				reporter: request.data.email,
+				reproduction: request.data.repro,
+				status: "open",
+					// bugBounty: {
+					// 	awarded: false,
+					// 	creditQuantity: 10,
+					// 	dateAwarded: null
+					// }
+			});
 
 		return send(messageOptions);
 	}
@@ -66,7 +86,7 @@ exports.handleSignup = onCall(
 		enforceAppCheck: true,
 	},
 	async (request) => {
-		if (request.auth === null) return { status: "failed", information: "User not signed in." };
+		if (!request.auth) return;
 		let currentWeekendDoc = await db.collection("activeWeekend").doc("default").get();
 		let currentWeekend = JSON.parse(currentWeekendDoc.data().information);
 		let attendeeRemoved = false;
@@ -239,6 +259,7 @@ exports.saveWeekendEvents = onCall(
 		enforceAppCheck: true,
 	},
 	(request) => {
+		if (!request.auth) return;
 		try {
 			let res = saveEvents(request);
 			return { status: "success", information: res };
@@ -253,6 +274,7 @@ exports.deleteEvent = onCall(
 		enforceAppCheck: true,
 	},
 	async (request) => {
+		if (!request.auth) return;
 		if (request.data.eventID) return await deleteCalendarEvent(request.data.eventID);
 		try {
 			for (let id of request.data.eventIDs) {
@@ -269,6 +291,7 @@ exports.createNewUser = onCall(
 		enforceAppCheck: true,
 	},
 	async (request) => {
+		if (!request.auth) return;
 		const usersDoc = await db.collection("settings").doc("config").get();
 		let usersData = usersDoc.data().data;
 		const userDoc = {
@@ -298,7 +321,9 @@ exports.createNewUser = onCall(
 	}
 );
 
-exports.updateWeekend = onSchedule("*/10 6-22 * 1-6,9-12 *", async (request) => {
+// At every 10th minute past every hour from 0 through 3 and every hour from 11 through 23 in every month from January through June and every month from September through December.
+//Adjusted for timezone
+exports.updateWeekend = onSchedule("*/10 0-3,11-23 * 1-6,9-12 *", async (request) => {
 	let queuedRef = await db.collection("activeWeekend").doc("queued").get();
 	let currentDate = new Date();
 	if (queuedRef.exists) {
@@ -383,7 +408,8 @@ exports.updateWeekend = onSchedule("*/10 6-22 * 1-6,9-12 *", async (request) => 
 							let data = await docRef.get();
 							data = data.data();
 							if (!data) continue;
-							data.credit += event.admission.credit || event.admission.credit === 0 ? event.admission.credit : 10;
+							data.credit +=
+								event.admission.credit || event.admission.credit === 0 ? event.admission.credit : 10;
 							let eventDate = new Date(activeWeekend.startDate + "T00:00:00");
 							eventDate.setDate(eventDate.getDate() + dayNum);
 							data.events.push({ title: event.title, date: eventDate.toDateString() });
@@ -393,7 +419,10 @@ exports.updateWeekend = onSchedule("*/10 6-22 * 1-6,9-12 *", async (request) => 
 							let data = await docRef.get();
 							if (!data.exists) continue;
 							data = data.data();
-							let creditLoss = event.admission.credit || event.admission.credit === 0 ? event.admission.credit / 2 | 0 : 5;
+							let creditLoss =
+								event.admission.credit || event.admission.credit === 0
+									? (event.admission.credit / 2) | 0
+									: 5;
 							data.credit = data.credit >= creditLoss ? data.credit - creditLoss : 0;
 							await docRef.set(data);
 						}
@@ -420,7 +449,7 @@ const updateUserInfoFunc = async () => {
 	if (!configDoc.dataSheet) return;
 	let sheet = await getSheetAsJSON(configDoc.dataSheet.ID);
 	if (sheet.status === "error") {
-		error("Error getting student info Google Sheet.");
+		// error("Error getting student info Google Sheet.");
 		return;
 	}
 	let studentInfoJSON = {};
@@ -433,7 +462,7 @@ const updateUserInfoFunc = async () => {
 	let adminDataSheet = (await db.collection("settings").doc("adminList").get()).data();
 	let adminSheet = await getSheetAsJSON(adminDataSheet.dataSheet.ID);
 	if (adminSheet.status === "error") {
-		error("Error getting admin info Google Sheet");
+		// error("Error getting admin info Google Sheet");
 		return;
 	}
 	let adminInfoJSON = {};
@@ -491,8 +520,8 @@ const updateUserInfoFunc = async () => {
 		return { status: "error", information: JSON.stringify(e) };
 	});
 };
-
-exports.updateUserInfoPeriodic = onSchedule("0 12 * 1-6,9-12 fri", async (request) => {
+// At 17:00 on Friday in every month from January through June and every month from September through December.
+exports.updateUserInfoPeriodic = onSchedule("0 17 * 1-6,9-12 fri", async (request) => {
 	// https://crontab.guru/#0_12_*_1-6,9-12_fri
 	await updateUserInfoFunc();
 });
@@ -502,6 +531,7 @@ exports.updateUserInfo = onCall(
 		enforceAppCheck: true,
 	},
 	async (request) => {
+		if (!request.auth) return;
 		await updateUserInfoFunc();
 	}
 );
@@ -511,6 +541,7 @@ exports.resetCredit = onCall(
 		enforceAppCheck: true,
 	},
 	async (request) => {
+		if (!request.auth) return;
 		let admin = await db.collection("admin").doc(request.auth.token.email).get();
 		if (!admin.exists) return { status: "error", information: "User not admin." };
 		let users = await db.collection("users").get();
@@ -522,7 +553,7 @@ exports.resetCredit = onCall(
 		});
 		try {
 			await batch.commit();
-			log(`Credit Reset by: ${request.auth.token.email}`);
+			// log(`Credit Reset by: ${request.auth.token.email}`);
 			return { status: "success", information: "All credit set to zero." };
 		} catch (error) {
 			return { status: "error", information: error.message };
@@ -535,6 +566,7 @@ exports.printRoster = onCall(
 		enforceAppCheck: true,
 	},
 	async (request) => {
+		if (!request.auth) return;
 		let activeWeekend = (await db.collection("activeWeekend").doc("default").get()).data();
 		if (!activeWeekend) return { status: "fail", information: "No active weekend found." };
 		activeWeekend = JSON.parse(activeWeekend.information);
@@ -600,14 +632,21 @@ exports.printRoster = onCall(
 
 		let messageOptions = emailOptions;
 		messageOptions.to = "pkkjx65dthv83@hpeprint.com";
-		messageOptions.cc = "polaris@westtown.edu";
-		messageOptions.subject = "Roster";
-		// messageOptions.text = request.data.text;
+		messageOptions.cc = `polaris@westtown.edu, ${request.auth.token.email}`;
+		messageOptions.subject = `Roster for ${event.title}`;
+		messageOptions.text = `${event.title}\n\n`
+
+		for (student of people) {
+			messageOptions.text += `${counter}. ${student.name} - ${student.email} - ${
+				student.cell ? student.cell : "no phone number"
+			} - ${student.grade}th\n`;
+		}
 		messageOptions.attachments = {
 			// path: "./testing.pdf",
 			filename: "Roster.pdf",
 			content: doc,
 		};
+		// console.log(messageOptions.text)
 
 		return send(messageOptions);
 	}
@@ -618,15 +657,32 @@ exports.manageAttendees = onCall(
 		enforceAppCheck: true, // Reject requests with missing or invalid App Check tokens.
 	},
 	async (request) => {
+		if (!request.auth) return;
 		let id = request.data.id;
 		let activeWeekend = (await db.collection("activeWeekend").doc("default").get()).data();
 		activeWeekend = JSON.parse(activeWeekend.information);
 		let event = activeWeekend.days[id[0]][id[1]];
+		if (request.data.addAttendee && event.admission.val === "none") {
+			let removed = false;
+			for (let i in event.signups) {
+				if (event.signups[i].email === request.auth.token.email) {
+					removed = true;
+					event.signups.splice(i, 1);
+				}
+			}
+			if (!removed) {
+				event.signups.push({ status: "approved", email: request.auth.token.email });
+			}
+			db.collection("activeWeekend")
+				.doc("default")
+				.set({ information: JSON.stringify(activeWeekend) });
+		}
 		return await manageAttendees(event);
 	}
 );
 
-exports.test = onCall(async () => {
+exports.test = onCall(async (request) => {
+	if (request.auth === null) return;
 	log("Running Test Function");
 	let current = new Date();
 	return { info: current.toString() };
